Finish 4.21

+++ part4/blog-list/controllers/blogs.js
@@ -18,7 +18,7 @@ blogsRouter.post('/', async (request, response) => {
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
 
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' });
+    return response.status(401).json({ error: 'Token invalid' });
   }
 
   const user = await User.findById(decodedToken.id);
@@ -42,6 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const requestedBlog = await Blog.findById(request.params.id);
+
+  if (!requestedBlog) {
+    return request.status(404).json({ error: 'Requested blog does not exist' });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'Token invalid' });
+  }
+
+  if (requestedBlog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({ error: 'User invalid' });
+  }
+
   await Blog.deleteOne({ _id: request.params.id });
 
   response.status(204).end();
